+++ spec/support/helpers.js
@@ -30,7 +30,7 @@ exports.runScenario = function(scenario, expectedResult) {
   });
 
   waitsFor(function() {
-    return typeof(result) != 'undefined';
+    return result !== undefined;
   }, "The scenario should have finished running.", 50);
 
   runs(function() {
@@ -38,4 +38,30 @@ exports.runScenario = function(scenario, expectedResult) {
   });
 
   return deferred.promise;
+};
+
+exports.addMatchers = function(jasmine) {
+  jasmine.addMatchers({
+    toBeAnError: function(expected) {
+
+      var actual = this.actual,
+          isNot = this.isNot,
+          classMatches = actual instanceof Error,
+          messageMatches = actual.message === expected;
+
+      this.message = function() {
+
+        var message = 'Expected an error with message "' + expected + '", got ';
+        if (!classMatches) {
+          message += typeof(actual);
+        } else if (!messageMatches) {
+          message += '"' + actual.message + '"';
+        }
+
+        return message;
+      };
+
+      return classMatches && messageMatches;
+    }
+  });
 };